refactor(notebook-cloud): move NotebookViewer into its own module
index.tsx is now a 120-line entry: module side-effects, the route switch in
App, CloudNotebookProviders, and the mount. The 1,000-line NotebookViewer and
its private helpers move to notebook-viewer.tsx. Pure code motion; typecheck
clean and the corpus guardrails stay green.
